// src/gatz/types.ts
export type SUUID = string;
export type SULID = string;
export type SDate = string;
// ...
export type UserProfile = {
  full_name: string | null;
  urls?: {
    website: string | null;
    twitter: string | null;
  };
};

export type LocationSettings = {
  enabled: boolean;
};

export type UserSettings = {
  notifications: NotificationSettings;
  location: LocationSettings;
};

export type Location = {
  id: SUUID,
  name: string,
}

export type NewLocationResponse = {
  location: Location,
  in_common: {
    friends: Contact["id"][],
    friends_of_friends: Contact["id"][],
  }
}

export type User = {
  id: SUUID;
  name: string;
  clock: HLC;
  avatar: string;
  phone_number: string;
  created_at: SDate;
  updated_at: SDate;
  is_test?: boolean;
  is_admin?: boolean;
  push_tokens?: { expo: PushToken };
  settings?: UserSettings;
  profile?: UserProfile;
};

type PushToken = {
  type: "expo";
  created_at: SDate;
  push_token: string;
};

export type HLC = {
  counter: number;
  node: SUUID;
  ts: SDate;
};
// ...
type BaseInviteLink = {
  id: SULID;
  code: string;
  expires_at: SDate;
  created_by: User["id"];
  created_at: SDate;
  used_at?: SDate;
  used_by?: User["id"];
};
export type GroupInviteLink = BaseInviteLink & {
  type: "group";
  group_id: Group["id"];
};

export type ContactInviteLink = BaseInviteLink & {
  type: "contact";
  contact_id: Contact["id"];
};

export type CrewInviteLink = BaseInviteLink & {
  type: "crew";
};

export type InviteLink = GroupInviteLink | ContactInviteLink | CrewInviteLink;

export type InviteLinkResponse =
  | {
    type: "group";
    invite_link: GroupInviteLink;
    group: Group;
    invited_by: Contact;
    in_common: {
      contact_ids: Contact["id"][];
      contacts: Contact[];
    };
  }
  | {
    type: "contact";
    invite_link: ContactInviteLink;
    contact: Contact;
    invited_by: Contact;
    in_common: {
      contact_ids: Contact["id"][];
      contacts: Contact[];
    };
  }
  | {
    type: "crew";
    invite_link: CrewInviteLink;
    invited_by: Contact;
    members: Contact[];
    group?: Group;
  };

export type ShareableInviteLink = {
  id: string;
  url: string;
  code: string;
};

// ======================================================================
// Group

export type GroupSettings = { 
  mode: "crew" | null; 
  member_mode: "open" | "closed" | "friends_of_friends" 
};

export type Group = {
  id: SULID;
  created_at: SDate;
  created_by: User["id"];
  updated_at: SDate;
  name: string;
  description?: string;
  avatar?: string;
  owner: User["id"];
  admins: User["id"][];
  members: User["id"][];
  joined_at: Record<User["id"], SDate>;
  archived_uids: User["id"][];
  settings: GroupSettings;
  is_public: boolean;
};
// ...
export type Contact = {
  id: SUUID;
  name: string;
  avatar: string | null;
  profile?: UserProfile;
};